Common module for preemptive cache
