+++ apps/transport/lib/transport/preemptive_api_cache.ex
@@ -3,37 +3,14 @@ defmodule Transport.PreemptiveAPICache do
   A module that populates the Cachex cache for the /api/datasets endpoint ("api-datasets-index")
   """
 
-  use GenServer
+  use Transport.PreemptiveBaseCache,
+    first_run: 0,
+    job_delay: :timer.seconds(300),
+    # more than twice job_delay to reduce the risk of parallel computation
+
+    cache_ttl: :timer.seconds(700)
+
   require Logger
-
-  @first_run 0
-  @job_delay :timer.seconds(300)
-  # slightly more than twice `@job_delay` to reduce the risk of parallel computation
-  @cache_ttl :timer.seconds(700)
-
-  def cache_ttl, do: @cache_ttl
-
-  def start_link(_opts) do
-    GenServer.start_link(__MODULE__, %{})
-  end
-
-  def init(state) do
-    # initial schedule is immediate, but via the same code path,
-    # to ensure we jump on the data
-    schedule_next_occurrence(@first_run)
-
-    {:ok, state}
-  end
-
-  def schedule_next_occurrence(delay \\ @job_delay) do
-    Process.send_after(self(), :tick, delay)
-  end
-
-  def handle_info(:tick, state) do
-    schedule_next_occurrence()
-    populate_cache()
-    {:noreply, state}
-  end
 
   def populate_cache do
     Logger.info("[preemptive-api-cache] Populating cache for /api/datasets…")
